create `mod target_specs` and move all related code there
