+++ crates/cargo-gpu/src/install.rs
@@ -1,12 +1,11 @@
 //! Install a dedicated per-shader crate that has the `rust-gpu` compiler in it.
 
-use crate::legacy_target_specs::write_legacy_target_specs;
 use crate::spirv_source::{
     get_channel_from_rustc_codegen_spirv_build_script, query_metadata, FindPackage as _,
 };
+use crate::target_specs::update_spec_files;
 use crate::{cache_dir, spirv_source::SpirvSource};
 use anyhow::Context as _;
-use cargo_metadata::Metadata;
 use spirv_builder::SpirvBuilder;
 use std::path::{Path, PathBuf};
 
@@ -196,83 +195,6 @@ package = "rustc_codegen_spirv"
         Ok(())
     }
 
-    /// Copy spec files from one dir to another, assuming no subdirectories
-    fn copy_spec_files(src: &Path, dst: &Path) -> anyhow::Result<()> {
-        std::fs::create_dir_all(dst)?;
-        let dir = std::fs::read_dir(src)?;
-        for dir_entry in dir {
-            let file = dir_entry?;
-            let file_path = file.path();
-            if file_path.is_file() {
-                std::fs::copy(file_path, dst.join(file.file_name()))?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Add the target spec files to the crate.
-    fn update_spec_files(
-        source: &SpirvSource,
-        install_dir: &Path,
-        dummy_metadata: &Metadata,
-        skip_rebuild: bool,
-    ) -> anyhow::Result<PathBuf> {
-        let mut target_specs_dst = install_dir.join("target-specs");
-        if !skip_rebuild {
-            if let Ok(target_specs) =
-                dummy_metadata.find_package("rustc_codegen_spirv-target-specs")
-            {
-                log::info!(
-                    "target-specs: found crate `rustc_codegen_spirv-target-specs` with manifest at `{}`", 
-                    target_specs.manifest_path
-                );
-
-                let target_specs_src = target_specs
-                    .manifest_path
-                    .as_std_path()
-                    .parent()
-                    .and_then(|root| {
-                        let src = root.join("target-specs");
-                        src.is_dir().then_some(src)
-                    })
-                    .context("Could not find `target-specs` directory within `rustc_codegen_spirv-target-specs` dependency")?;
-                if source.is_path() {
-                    // skip copy
-                    log::info!(
-                        "target-specs: source is local path, use target-specs from `{}`",
-                        target_specs_src.display()
-                    );
-                    target_specs_dst = target_specs_src;
-                } else {
-                    // copy over the target-specs
-                    log::info!(
-                        "target-specs: Copy target specs from `{}`",
-                        target_specs_src.display()
-                    );
-                    Self::copy_spec_files(&target_specs_src, &target_specs_dst)
-                        .context("copying target-specs json files")?;
-                }
-            } else {
-                // use legacy target specs bundled with cargo gpu
-                if source.is_path() {
-                    // This is a stupid situation:
-                    // * We can't be certain that there are `target-specs` in the local checkout (there may be some in `spirv-builder`)
-                    // * We can't dump our legacy ones into the `install_dir`, as that would modify the local rust-gpu checkout
-                    // -> do what the old cargo gpu did, one global dir for all target specs
-                    // and hope parallel runs don't shred each other
-                    target_specs_dst = cache_dir()?.join("legacy-target-specs-for-local-checkout");
-                }
-                log::info!(
-                    "target-specs: Writing legacy target specs to `{}`",
-                    target_specs_dst.display()
-                );
-                write_legacy_target_specs(&target_specs_dst)?;
-            }
-        }
-
-        Ok(target_specs_dst)
-    }
-
     /// Install the binary pair and return the [`InstalledBackend`], from which you can create [`SpirvBuilder`] instances.
     ///
     /// # Errors
@@ -339,7 +261,7 @@ package = "rustc_codegen_spirv"
 
         log::debug!("update_spec_files");
         let target_spec_dir =
-            Self::update_spec_files(&source, &install_dir, &dummy_metadata, skip_rebuild)
+            update_spec_files(&source, &install_dir, &dummy_metadata, skip_rebuild)
                 .context("writing target spec files")?;
 
         if !skip_rebuild {
